파티클추가, 비내리기, 폭발진행중
형변환, 오류뜨는거 일단 안뜨게 바꿔놨는데 이건 머지 안해도 상관없음

+++ Engine/Private/Graphic_Device.cpp
@@ -119,7 +119,7 @@ HRESULT CGraphic_Device::Draw_Font(const _wstring& str, LPRECT pRect)
 	m_pSprite->Begin(D3DXSPRITE_ALPHABLEND);
 
 	if (FAILED(m_pFont->DrawText(m_pSprite,
-		str.c_str(), str.length(), pRect,
+		str.c_str(), (int)str.length(), pRect,
 		DT_LEFT | DT_VCENTER | DT_WORDBREAK,
 		D3DCOLOR_ARGB(255, 0, 0, 0))))
 		return E_FAIL;
